requirements.txt
streamlit>=1.28.0
PyPDF2>=3.0.0
openai>=1.3.0
requests>=2.31.0
python-dateutil>=2.8.2
google-generativeai>=0.3.0
pytest>=7.4.0
pytest-mock>=3.11.0
python-dotenv>=1.0.0

